Make a clearer delineation between true unit tests and integration tests
